remove the old evaluator
