--- src/compiler/api/GF/Term.hs
@@ -1,12 +1,0 @@
-module GF.Term (renameSourceTerm,
-                Globals(..), ConstValue(..), EvalM, stdPredef,
-                Value(..), showValue, Thunk, newThunk, newEvaluatedThunk,
-                evalError, evalWarn,
-                inferLType, inferLType', checkLType, checkLType',
-                normalForm, normalFlatForm, normalStringForm,
-                unsafeIOToEvalM, force
-               ) where
-
-import GF.Compile.Rename
-import GF.Compile.Compute.Concrete
-import GF.Compile.TypeCheck.Concrete
